ml-service/requirements.txt
fastapi==0.111.0
uvicorn==0.27.1
pydantic==2.7.1
pandas==2.2.3
numpy==2.2.5
scikit-learn==1.6.1
joblib==1.4.2
python-multipart==0.0.7
python-dotenv==1.1.0
httpx==0.27.0
xgboost==2.0.3